moved the moveSection to the draggable container component to simplify the template components
